pulled the image picker logic out; #13
