+++ MemeMe/MemeEditorViewController+Extension.swift
@@ -1,0 +1,30 @@
+//
+//  MemeEditorViewController+Extension.swift
+//  MemeMe
+//
+//  Created by Shane Richards on 4/19/20.
+//  Copyright © 2020 Shane Richards. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+// image picker logic
+
+extension MemeEditorViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
+    
+    
+    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
+        if let image = info[.originalImage] as? UIImage {
+            imageView.image = image
+            shareButton.isEnabled = true
+        }
+        
+        dismiss(animated: true, completion: nil)
+    }
+    
+    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+        dismiss(animated: true, completion: nil)
+    }
+    
+}
